books wrap based on screen size and are fully responsive :book::dog2:
